requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
Crawl4AI==0.4.247
cryptography==44.0.0
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
eval_type_backport==0.2.0
executing==2.1.0
fake-http-header==0.3.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
gotrue==2.11.1
greenlet==3.1.1
griffe==1.5.1
groq==0.13.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.27.1
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpath-python==1.0.6
jsonpickle==0.9.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.57.8
logfire==2.6.2
logfire-api==2.6.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
mistralai==1.5.0
mockito==1.5.3
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.57.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
playwright==1.49.1
plotly==6.0.0
pluggy==1.5.0
postgrest==0.19.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.1
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic-ai==0.0.17
pydantic-ai-slim==0.0.17
pydantic_core==2.27.1
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyperclip==1.9.0
pytest==8.3.4
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-weather==2.0.7
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
realtime==2.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rizaio==0.7.0
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
st-copy-to-clipboard==0.1.6
storage3==0.11.0
streamlit==1.40.2
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
tenacity==9.0.0
tf-playwright-stealth==1.1.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
weather-api==1.0
websockets==13.1
wikipedia==1.4.0
Wikipedia-API==0.8.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
